message_fetch: Move parse_anchor_value to zerver.lib.narrow.
